requirements.txt
torch
torchvision
torchaudio
wandb==0.23.1
numpy
pandas
pillow
matplotlib
seaborn
scikit-learn
kagglehub
timm
ipykernel
tqdm